+++ crates/indexer-client/src/models/hub_lp_allowlist.rs
@@ -19,9 +19,9 @@ pub struct HubLpAllowlist {
     pub lp: Option<String>,
     /// Event sequence at which this allowlist entry became current  Note: This is a Primary Key.<pk/>
     #[serde(rename = "valid_from_seq", skip_serializing_if = "Option::is_none")]
-    pub valid_from_seq: Option<i32>,
+    pub valid_from_seq: Option<i64>,
     #[serde(rename = "valid_to_seq", skip_serializing_if = "Option::is_none")]
-    pub valid_to_seq: Option<i32>,
+    pub valid_to_seq: Option<i64>,
     /// Whether this LP may deposit into the fast-fill vault (withdrawals are always allowed)
     #[serde(rename = "allowed", skip_serializing_if = "Option::is_none")]
     pub allowed: Option<bool>,
